# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/change/password": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This endpoint is used to change password in admin side.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "change password admin",
                "parameters": [
                    {
                        "description": "Change password data",
                        "name": "ChangePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.ChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfuly change password",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing data",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponsAuthorization"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
        "/admin/owner/register/employe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This endpoint is used to register new employee to access admin side.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "signup for admin side",
                "parameters": [
                    {
                        "description": "Signup data",
                        "name": "Signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.SignupEmploye"
                        }
                    },
                    {
                        "enum": [
                            "Tetap",
                            "Kontrak",
                            "Freelance"
                        ],
                        "type": "string",
                        "description": "Employment Type (Tetap, Kontrak, Freelance)",
                        "name": "employementType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "Man",
                            "Woman"
                        ],
                        "type": "string",
                        "description": "Gender (Man, Woman)",
                        "name": "gender",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "Staff",
                            "Owner"
                        ],
                        "type": "string",
                        "description": "Position (Staff, Owner)",
                        "name": "position",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfuly register employee",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing data",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponsAuthorization"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
...
        "/auth/login/admin": {
            "post": {
                "description": "This endpoint is used to login by email and password.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Admin login",
                "parameters": [
                    {
                        "description": "login data",
                        "name": "Login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user successfully login",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponsAuthorization"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
        "/auth/login/user": {
            "post": {
                "description": "This enpoint is used to authenticate users with send cridential data (email dan password).",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login Input",
                        "name": "Login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfuly login, return a token to access enpoint for user",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Request invalid or the data sent is incorrect",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorBadRequest"
                        }
                    },
                    "500": {
                        "description": "Mistake in the server",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
        "/auth/signup/user": {
            "post": {
                "description": "This endpoint is used to create a new user account by submitting registration data in JSON format. Users are required to provide information email, and password. The data will be validated on the server side before the account is created temporary, client have navigate to endpoint signup verification on the client side.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Create New User Account",
                "parameters": [
                    {
                        "description": "User Registration Data (name, email, password)",
                        "name": "Signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account successfully created temporery, return a token to verification email",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid request or incomplete data",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorBadRequest"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
        "/auth/signup/user/verification": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This endpoint is used to verify the user's email during the registration process. A verification code is sent to the user's email, and the user must input this code to complete the signup process successfully. The request must include the verification code in the form body.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Verify User Email for Signup",
                "parameters": [
                    {
                        "description": "Email Verification Data (code)",
                        "name": "Signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.Verification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email successfully verified, account registration completed",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing verification code",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponsAuthorization"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
...
        "/user/change/password": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This endpoint is used to change password in client side or user.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "change password user",
                "parameters": [
                    {
                        "description": "Change password data",
                        "name": "ChangePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.ChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfuly change password",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing data",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponsAuthorization"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/authentication.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "authentication.ChangePassword": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "passwordBefore": {
                    "type": "string"
                }
            }
        },
        "authentication.RegisterUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "authentication.ResponsAuthorization": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "authentication.ResponseErrorBadRequest": {
            "type": "object",
            "properties": {
                "ErrorField": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "authentication.ResponseErrorInternalServer": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "authentication.ResponseSuccess": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "string"
                }
            }
        },
        "authentication.SignupEmploye": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employementType": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "whatsapp": {
                    "type": "string"
                }
            }
        },
        "authentication.Verification": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                }
            }
        },
...
    }
}